+++ internal/usecase/forum/model.go
@@ -1,7 +1,7 @@
 package forum
 
 type Forum struct {
-	Posts   *int32  `json:"posts"`
+	Posts   *int64  `json:"posts"`
 	Slug    *string `json:"slug"`
 	Threads *int32  `json:"threads"`
 	Title   *string `json:"title"`
